+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Add #PAC-34
 * Add #PAC-85
+* Add techdivision/import#175
 
 # Version 15.2.1